# se-ex-client/target/classes/api.yaml
{"openapi":"3.0.1","info":{"title":"Serasa Experian","version":"0"},"servers":[{"url":"http://localhost:8080/api"}],"tags":[{"name":"score"},{"name":"afinidade"},{"name":"pessoa"}],"paths":{"/score":{"get":{"tags":["score"],"summary":"Lista scores","description":"Retorna uma lista de todas os scores cadastrados","operationId":"listAllScores","responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Score"}}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Score"}}}}}}},"post":{"tags":["score"],"summary":"Cria score","operationId":"saveController","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScoreCommand"}}},"required":true},"responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Score"}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Score"}}}}}}},"/pessoa":{"get":{"tags":["pessoa"],"summary":"Lista pessoas","description":"Retorna uma lista de todas as pessoas cadastradas","operationId":"listAllPessoas","responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PessoaInListDto"}}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PessoaInListDto"}}}}}}},"post":{"tags":["pessoa"],"summary":"Cria pessoa","operationId":"savePessoa","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaCommand"}}},"required":true},"responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pessoa"}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pessoa"}}}}}}},"/afinidade":{"get":{"tags":["afinidade"],"summary":"Lista afinidade","description":"Retorna uma lista de todas as afinidades cadastradas","operationId":"listAllAfinidades","responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Afinidade"}}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Afinidade"}}}}}}},"post":{"tags":["afinidade"],"summary":"Cria afinidade","operationId":"saveAfinidade","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Afinidade"}}},"required":true},"responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Afinidade"}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Afinidade"}}}}}}},"/pessoa/{id}":{"get":{"tags":["pessoa"],"summary":"Lista pessoas","description":"Retorna uma lista de todas as pessoas cadastradas","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"204":{"description":"No Content","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SingularPessoaDto"}}}},"201":{"description":"Created","content":{"application/json":{}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Ok","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SingularPessoaDto"}}}}}}}},"components":{"schemas":{"ScoreCommand":{"type":"object","properties":{"scoreDescricao":{"type":"string"},"inicial":{"minimum":0,"type":"integer","format":"int32"},"ffinal":{"maximum":1000,"type":"integer","format":"int32"}}},"ApiErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"Score":{"type":"object","properties":{"descricao":{"type":"string"},"inicial":{"minimum":0,"type":"integer","format":"int32"},"ffinal":{"maximum":1000,"type":"integer","format":"int32"}}},"PessoaCommand":{"type":"object","properties":{"nome":{"type":"string"},"telefone":{"type":"string"},"cidade":{"type":"string"},"estado":{"type":"string","enum":["AC","AL","AP","AM","BA","CE","DF","ES","GO","MA","MT","MS","MG","PA","PB","PR","PE","PI","RJ","RN","RS","RO","RR","SC","SP","SE","TO"]},"regiao":{"type":"string"},"idade":{"type":"integer","format":"int32"},"score":{"type":"integer","format":"int32"}}},"Afinidade":{"type":"object","properties":{"regiao":{"type":"string"},"estados":{"type":"array","items":{"type":"string","enum":["AC","AL","AP","AM","BA","CE","DF","ES","GO","MA","MT","MS","MG","PA","PB","PR","PE","PI","RJ","RN","RS","RO","RR","SC","SP","SE","TO"]}}}},"Pessoa":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dataInclusao":{"type":"string","format":"date-time"},"nome":{"type":"string"},"telefone":{"type":"string"},"cidade":{"type":"string"},"estado":{"type":"string","enum":["AC","AL","AP","AM","BA","CE","DF","ES","GO","MA","MT","MS","MG","PA","PB","PR","PE","PI","RJ","RN","RS","RO","RR","SC","SP","SE","TO"]},"regiao":{"$ref":"#/components/schemas/Afinidade"},"idade":{"type":"integer","format":"int32"},"score":{"$ref":"#/components/schemas/Score"}}},"PessoaInListDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"scoreDescricao":{"type":"string"},"estados":{"type":"array","items":{"type":"string","enum":["AC","AL","AP","AM","BA","CE","DF","ES","GO","MA","MT","MS","MG","PA","PB","PR","PE","PI","RJ","RN","RS","RO","RR","SC","SP","SE","TO"]}},"idade":{"type":"integer","format":"int32"}}},"SingularPessoaDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"scoreDescricao":{"type":"string"},"estados":{"type":"array","items":{"type":"string","enum":["AC","AL","AP","AM","BA","CE","DF","ES","GO","MA","MT","MS","MG","PA","PB","PR","PE","PI","RJ","RN","RS","RO","RR","SC","SP","SE","TO"]}},"idade":{"type":"integer","format":"int32"},"telefone":{"type":"string"}}}}}}