# se-ex-server/src/main/java/com/github/hyagosouzza/seexserver/dtos/SingularPessoaDto.java
package com.github.hyagosouzza.seexserver.dtos;

public class SingularPessoaDto extends PessoaDto {

	private String telefone;

	public String getTelefone() {
		return telefone;
	}

	public void setTelefone(String telefone) {
		this.telefone = telefone;
	}

}